created TrophyCabinet component
